Some adjustments to make nimx compile with latest opengl head.

+++ nimx/font.nim
@@ -297,7 +297,7 @@ proc bakeChars(f: Font, start: int32): CharInfo =
         when dumpDebugBitmaps:
             discard stbi_write_bmp("atlas_nimx_df_" & $fSize & "_" & $start & "_" & $width & "x" & $height & ".bmp", width, height, 1, addr temp_bitmap[0])
         let gl = result.prepareTexture()
-        gl.texImage2D(gl.TEXTURE_2D, 0, gl.ALPHA, width, height, 0, gl.ALPHA, gl.UNSIGNED_BYTE, addr temp_bitmap[0])
+        gl.texImage2D(gl.TEXTURE_2D, 0, GLint(gl.ALPHA), width, height, 0, gl.ALPHA, gl.UNSIGNED_BYTE, addr temp_bitmap[0])
 
     gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
     gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
